requirements.txt
opencv-python~=4.10.0.84
torch~=1.12.1+cu113
numpy~=1.24.4
pytorch3d~=0.7.1
dataclasses~=0.6
packaging~=24.1
omegaconf~=2.3.0
imageio~=2.27.0
matplotlib~=3.9.2
pillow~=9.4.0
jaxtyping~=0.2.34
typeguard~=2.13.3
open3d~=0.18.0
einops~=0.7.0
setuptools~=59.5.0
transformers~=4.34.1
plyfile~=1.1
diffusers~=0.19.3
torchvision~=0.13.1+cu113
tqdm~=4.66.5
gradio~=4.8.0
pytest~=6.2.5
scipy~=1.13.1
rembg~=2.0.59
accelerate~=0.34.0
datasets~=2.21.0
pandas~=2.2.2
seaborn~=0.13.2
lpips~=0.1.4
scikit-image~=0.22.0